+++ CHANGELOG.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 - Commit: [`HEAD`](https://github.com/aviatesk/JETLS.jl/commit/HEAD)
-- Diff: [`5643648...HEAD`](https://github.com/aviatesk/JETLS.jl/compare/5643648...HEAD)
+- Diff: [`d15f92f...HEAD`](https://github.com/aviatesk/JETLS.jl/compare/d15f92f...HEAD)
 
 ### Announcement
 
@@ -43,6 +43,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 > ```
 > This disables analysis for matched files. Basic features like completion still might work, but most LSP features will be unfunctional.
 > Note that `analysis_overrides` is provided as a temporary workaround and may be removed or changed at any time. A proper fix is being worked on.
+
+### Fixed
+
+- Fixed spurious `Any` type annotations caused by free static parameters left inside inferred method argument types, such as `Vector{T}` in `f(a::Vector{T}) where {T}`. (Closed https://github.com/aviatesk/JETLS.jl/issues/768)
+
+- Fixed type annotation for script-mode files so inferred `const` globals use their actual type instead of an internal `JET.AbstractBindingState`, avoiding spurious annotations.
+
+## 2026-06-20
+
+- Commit: [`d15f92f`](https://github.com/aviatesk/JETLS.jl/commit/d15f92f)
+- Diff: [`5643648...d15f92f`](https://github.com/aviatesk/JETLS.jl/compare/5643648...d15f92f)
+- Installation:
+  ```bash
+  julia -e 'using Pkg; Pkg.Apps.add(; url="https://github.com/aviatesk/JETLS.jl", rev="2026-06-20")'
+  ```
 
 ### Fixed